+++ package.json
@@ -20,6 +20,8 @@
   "dependencies": {
     "body-parser": "^1.15.2",
     "express": "^4.14.0",
+    "express-handlebars": "^3.0.0",
+    "handlebars": "^4.0.5",
     "hbs": "^4.0.1",
     "mongoose": "^4.6.6"
   }
