+++ views/index.hbs
@@ -1,0 +1,1 @@
+<h1>yo</h1>
