+++ views/layout.hbs
@@ -1,0 +1,11 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <meta charset="utf-8">
+    <title>URL Shortner</title>
+  </head>
+  <body>
+    <h1>URL Shorten</h1>
+    {{{body}}}
+  </body>
+</html>
