+++ views/welcome.hbs
@@ -1,0 +1,1 @@
+<h1>Welcome</h1>
